Remove selfcheck.py, keep random_map in map.py
The no-answer-key consistency checks are dropped; random_map (random maps
with the goal guaranteed reachable) moves verbatim into map.py, where the
practice-map regeneration and students can keep using it. Verified: the
practice maps regenerate byte-identically through the relocated function.
